NEEDS REVIEW: attempt to change write frequency to yearly...
The setting KPRNT is the first number in settings.txt, and should take
care of this easily. However, there is some special logic to force the
program to write the first year, and this breaks the array indexing if
KPRNT == 1. I have tried to fix this by adding a special case in two
places in the code. It runs, but the output file is suspiciouly large
compared with the KPRNT == 2 case.